Separate implementations of Redis but prefer `ioredis` - First use `redis` - Latter use `ioredis` => preferable
