+++ src/app.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, HttpStatus, Inject, Redirect } from '@nestjs/common';
 import { AppService } from './app.service';
-import { RedisService12 } from './building-blocks/infra/redis/redis12.service';
 import { RedisService } from './building-blocks/infra/redis/redis.service';
+import { RedisIoRedisService } from './building-blocks/infra/redis-ioredis/redis-ioredis.service';
 import { ApiExcludeEndpoint, ApiTags } from '@nestjs/swagger';
 
 @ApiTags("App")
@@ -9,8 +9,8 @@ import { ApiExcludeEndpoint, ApiTags } from '@nestjs/swagger';
 export class AppController {
   constructor(
     private readonly appService: AppService,
-    private readonly _redisService: RedisService,
-    private readonly _redisService12: RedisService12
+    private readonly _redisIoRedisService: RedisIoRedisService,
+    private readonly _redisService: RedisService
   ) {}
 
   @Get("")
@@ -23,14 +23,14 @@ export class AppController {
     return this.appService.getHello();
   }
 
+  @Get('redis-ioredis/ping')
+  async pingRedisIoRedis(): Promise<string> {
+    return await this._redisIoRedisService.ping();
+  }
+
   @Get('redis/ping')
   async pingRedis(): Promise<string> {
     return await this._redisService.ping();
-  }
-
-  @Get('redis12/ping')
-  async pingRedis12(): Promise<string> {
-    return await this._redisService12.ping();
   }
   
 }
