+++ src/building-blocks/infra/caching/caching.module.ts
@@ -2,7 +2,7 @@ import { CacheModule } from '@nestjs/cache-manager';
 import { DynamicModule, Global, Module } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { redisStore } from 'cache-manager-ioredis-yet';
-import { RedisService } from '../redis/redis.service';
+import { RedisIoRedisService } from '../redis-ioredis/redis-ioredis.service';
 import { CachingProvider } from './caching.provider';
 
 @Global()
@@ -12,8 +12,8 @@ export class CachingModule {
 
         const cacheDynamicModule: DynamicModule = CacheModule.registerAsync({
             isGlobal: true,
-            inject: [ConfigService, RedisService],
-            useFactory: async (configService: ConfigService, redisService: RedisService) => {
+            inject: [ConfigService, RedisIoRedisService],
+            useFactory: async (configService: ConfigService, redisService: RedisIoRedisService) => {
 
                 const ttl = configService.get<number>('CACHE_TTL') || parseInt(process.env.CACHE_TTL!, 10) || 6000;
