+++ test/app.e2e.ts
@@ -3,15 +3,16 @@ import * as request from 'supertest';
 import { httpServer } from '@test/test.setup';
 
 describe('AppController (e2e)', () => {
-    test("/redis/ping (GET)", async () => {
-        const response = await request(httpServer).get("/redis/ping");
+
+    test("/redis-ioredis/ping (GET)", async () => {
+        const response = await request(httpServer).get("/redis-ioredis/ping");
 
         expect(response.status).toBe(HttpStatus.OK);
         expect(response.text).toBeDefined();
     });
 
-    test("/redis12/ping", async () => {
-        const response = await request(httpServer).get("/redis12/ping");
+    test("/redis/ping", async () => {
+        const response = await request(httpServer).get("/redis/ping");
 
         expect(response.status).toBe(HttpStatus.OK);
         expect(response.text).toBeDefined();
