+++ test/test-infra/redis/redis.e2e-spec.ts
@@ -2,8 +2,10 @@ import { faker } from "@faker-js/faker";
 import { RedisService } from "@src/building-blocks/infra/redis/redis.service";
 import { app } from "@test/test.setup";
 
-describe("Test for Redis", () => {
-    let redisService: RedisService | undefined;
+
+
+describe("Test Redis", () => {
+    let redisService: RedisService;
 
     const cacheKey = faker.string.uuid();
     const cacheObjects = faker.helpers.multiple(() => ({
@@ -12,13 +14,13 @@ describe("Test for Redis", () => {
     }), { count: 5 });
     
     beforeEach(async () => {
-        redisService = app.get(RedisService, { strict: false });
+        redisService = app.get<RedisService>(RedisService);
         if (redisService.getRedisClient()) {
             await redisService.set(cacheKey, cacheObjects);
         }
     });
 
-    it("RedisService", async () => {
+    it(`${RedisService.name}`, async () => {
         if (redisService.getRedisClient()) {
             const objectsFromCache = await redisService.get(cacheKey);
             expect(objectsFromCache).toEqual(cacheObjects);
