+++ tsconfig.json
@@ -20,7 +20,6 @@
     "allowJs": true,
     "paths": {
       "@src/*": ["src/*"],
-      "@infrastructures/*": ["src/infrastructures/*"],
       "@building-blocks/*": ["src/building-blocks/*"],
       "@projects/*": ["src/projects/*"],
       "@test/*": ["test/*"],
